added new pages - Personal Page, Size Page, Package Page and others, added interceptor on 401 status code, and handler on 403 status code

+++ StockAdmin/Scripts/Server/HttpHandler.cs
@@ -1,7 +1,10 @@
 ﻿using System.Collections.Generic;
+using System.Net;
 using System.Net.Http;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
+using StockAdmin.Models;
+using StockAdmin.Scripts.Repositories;
 
 namespace StockAdmin.Scripts.Server;
 
@@ -11,24 +14,53 @@ public class HttpHandler<TEntity>
     {
         HttpClient httpClient = new HttpClient();
         httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
-        return await httpClient.GetFromJsonAsync<List<TEntity>>($"{ServerConstants.ServerAddress}{point}");
+        var responseMessage = await httpClient.GetAsync($"{ServerConstants.ServerAddress}{point}");
+        if (responseMessage.IsSuccessStatusCode == false)
+        {
+            await RefreshToken(responseMessage);
+            httpClient.DefaultRequestHeaders.Remove("Authorization");
+            httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
+            responseMessage = await httpClient.GetAsync($"{ServerConstants.ServerAddress}{point}");
+        }
+        return await responseMessage.Content.ReadFromJsonAsync<List<TEntity>>();
     }
     public async Task<TEntity?> GetFromJsonAsync(string point)
     {
         HttpClient httpClient = new HttpClient();
         httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
-        return await httpClient.GetFromJsonAsync<TEntity>($"{ServerConstants.ServerAddress}{point}");
+
+        var responseMessage = await httpClient.GetAsync($"{ServerConstants.ServerAddress}{point}");
+        if (responseMessage.IsSuccessStatusCode == false)
+        {
+            await RefreshToken(responseMessage);
+            httpClient.DefaultRequestHeaders.Remove("Authorization");
+            httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
+            responseMessage = await httpClient.GetAsync($"{ServerConstants.ServerAddress}{point}");
+        }
+        return await responseMessage.Content.ReadFromJsonAsync<TEntity>();
     }
     
     public async Task<TEntity?> PostAsJsonAsync(string point, TEntity entity)
     {
         HttpClient httpClient = new HttpClient();
+        
         if (ServerConstants.Token != null)
         {
             httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
         }
+
+
         HttpResponseMessage responseMessage =
             await httpClient.PostAsJsonAsync($"{ServerConstants.ServerAddress}{point}", entity);
+        if (responseMessage.IsSuccessStatusCode == false)
+        {
+            await RefreshToken(responseMessage);
+            httpClient.DefaultRequestHeaders.Remove("Authorization");
+            httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
+            responseMessage = 
+                await httpClient.PostAsJsonAsync($"{ServerConstants.ServerAddress}{point}", entity);
+            
+        }
         return await responseMessage.Content.ReadFromJsonAsync<TEntity>();
     }
 
@@ -36,8 +68,18 @@ public class HttpHandler<TEntity>
     {
         HttpClient httpClient = new HttpClient();
         httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
-        HttpResponseMessage responseMessage =
-            await httpClient.PutAsJsonAsync($"{ServerConstants.ServerAddress}{point}", entity);
+
+        var responseMessage = await httpClient.PutAsJsonAsync($"{ServerConstants.ServerAddress}{point}", entity);
+        
+        if (responseMessage.IsSuccessStatusCode == false)
+        {
+            await RefreshToken(responseMessage);
+            httpClient.DefaultRequestHeaders.Remove("Authorization");
+            httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
+            responseMessage =
+                await httpClient.PutAsJsonAsync($"{ServerConstants.ServerAddress}{point}", entity);
+        }
+
         return await responseMessage.Content.ReadFromJsonAsync<TEntity>();
     }
     
@@ -45,6 +87,26 @@ public class HttpHandler<TEntity>
     {
         HttpClient httpClient = new HttpClient();
         httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
-        await httpClient.DeleteAsync($"{ServerConstants.ServerAddress}{point}");
+
+        var responseMessage = await httpClient.DeleteAsync($"{ServerConstants.ServerAddress}{point}");
+        if (responseMessage.IsSuccessStatusCode == false)
+        {
+            await RefreshToken(responseMessage);
+            httpClient.DefaultRequestHeaders.Remove("Authorization");
+            httpClient.DefaultRequestHeaders.Add("Authorization", $"Bearer {ServerConstants.Token.token}");
+            await httpClient.DeleteAsync($"{ServerConstants.ServerAddress}{point}");
+        }
+
+
+    }
+
+    private async Task RefreshToken(HttpResponseMessage responseMessage)
+    {
+        if (responseMessage.StatusCode == HttpStatusCode.Unauthorized)
+        {
+            var repository = new PersonRepository();
+            Auth? auth = await repository.LoginAsync(new Person() { email = ServerConstants.Login, password = ServerConstants.Password });
+            ServerConstants.Token = auth!;
+        }
     }
 }
